+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to
 
 ### Fixed
 
+- 🐛(backend) skip session creation for the liveness probe
 - 🐛(frontend) preserve page titles when adding an emoji #2586
 
 ## [v5.6.1] - 2026-09-04